M  Cargo.toml, M  run.py, M  src/board.rs, M  src/chess/castle.rs, M  src/chess/move_generator.rs, M  src/chess/sub_board.rs, M  src/chess/sub_board_builder.rs, M  src/constants.rs, M  src/evaluate.rs, M  src/lib.rs, M  src/search.rs, M  src/sort.rs, M  src/tests.rs, M  src/useful_macros.rs, M  src/utils/bitboard.rs, M  src/utils/files.rs, D  src/utils/hash_utils.rs, M  src/utils/mod.rs, M  src/utils/pv_utils.rs, M  src/utils/ranks.rs, M  src/utils/square.rs, M  tests/board.rs

+++ src/useful_macros.rs
@@ -1,4 +1,4 @@
-#[cfg(debug_assertions)]
+#[cfg(not(feature = "speed"))]
 #[macro_export]
 macro_rules! get_item_unchecked {
     ($vec:expr, $index:expr) => {
@@ -12,7 +12,7 @@ macro_rules! get_item_unchecked {
     };
 }
 
-#[cfg(debug_assertions)]
+#[cfg(not(feature = "speed"))]
 #[macro_export]
 macro_rules! get_item_unchecked_mut {
     ($vec:expr, $index:expr) => {
@@ -26,7 +26,7 @@ macro_rules! get_item_unchecked_mut {
     };
 }
 
-#[cfg(not(debug_assertions))]
+#[cfg(feature = "speed")]
 #[macro_export]
 macro_rules! get_item_unchecked {
     ($vec:expr, $index:expr) => {
@@ -44,7 +44,7 @@ macro_rules! get_item_unchecked {
     };
 }
 
-#[cfg(not(debug_assertions))]
+#[cfg(feature = "speed")]
 #[macro_export]
 macro_rules! get_item_unchecked_mut {
     ($vec:expr, $index:expr) => {
@@ -80,9 +80,9 @@ macro_rules! inverse_interpolate {
 macro_rules! match_interpolate {
     ($new_start:expr, $new_end:expr, $old_start:expr, $old_end:expr, $old_value:expr) => {
         interpolate!(
-            ($new_start) as f64,
-            ($new_end) as f64,
-            inverse_interpolate!(($old_start) as f64, ($old_end) as f64, ($old_value) as f64)
+            $new_start,
+            $new_end,
+            inverse_interpolate!($old_start, $old_end, $old_value)
         )
     };
 }
